+++ database/migrations/2020_08_07_105331_create_booking_installment_paids_table.php
@@ -15,9 +15,9 @@ class CreateBookingInstallmentPaidsTable extends Migration
     {
         Schema::create('booking_installment_paids', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('booking_id');
-            $table->unsignedInteger('booking_payment_through_id');
-            $table->unsignedInteger('created_by');
+            $table->unsignedBigInteger('booking_id');
+            $table->unsignedBigInteger('booking_payment_through_id');
+            $table->unsignedBigInteger('created_by');
             $table->decimal('amount',17,2)->comment('paid amount');
             $table->index('booking_id');
             $table->index('booking_payment_through_id');
